Extract _plex_cache.py from _legacy.py
